rusk: rename `chain` into `node`
